requirements.txt
kelvin-python-sdk[ai]